chore(ui): move components folder to lib/matador
