--- app/components/Matador/SideBar.tsx
@@ -1,109 +1,0 @@
-import { Dialog, Transition } from "@headlessui/react";
-import React from "react";
-
-type SideBarProps = React.PropsWithChildren<{
-  srcImage: string;
-  name: string;
-}>;
-
-export function SideBar({ srcImage, name, children }: SideBarProps) {
-  const [sidebarOpen, setSidebarOpen] = React.useState(false);
-  return (
-    <>
-      <Transition.Root show={sidebarOpen} as={React.Fragment}>
-        <Dialog
-          as="div"
-          className="fixed inset-0 flex z-40 md:hidden"
-          onClose={setSidebarOpen}
-        >
-          <Transition.Child
-            as={React.Fragment}
-            enter="transition-opacity ease-linear duration-300"
-            enterFrom="opacity-0"
-            enterTo="opacity-100"
-            leave="transition-opacity ease-linear duration-300"
-            leaveFrom="opacity-100"
-            leaveTo="opacity-0"
-          >
-            <Dialog.Overlay className="fixed inset-0 bg-gray-600 bg-opacity-75" />
-          </Transition.Child>
-          <Transition.Child
-            as={React.Fragment}
-            enter="transition ease-in-out duration-300 transform"
-            enterFrom="-translate-x-full"
-            enterTo="translate-x-0"
-            leave="transition ease-in-out duration-300 transform"
-            leaveFrom="translate-x-0"
-            leaveTo="-translate-x-full"
-          >
-            <div className="relative flex-1 flex flex-col max-w-xs w-full bg-gray-800">
-              <Transition.Child
-                as={React.Fragment}
-                enter="ease-in-out duration-300"
-                enterFrom="opacity-0"
-                enterTo="opacity-100"
-                leave="ease-in-out duration-300"
-                leaveFrom="opacity-100"
-                leaveTo="opacity-0"
-              >
-                <div className="absolute top-0 right-0 -mr-12 pt-2">
-                  <button
-                    type="button"
-                    className="ml-1 flex items-center justify-center h-10 w-10 rounded-full focus:outline-none focus:ring-2 focus:ring-inset focus:ring-white"
-                    onClick={() => setSidebarOpen(false)}
-                  >
-                    <span className="sr-only">Close sidebar</span>
-                  </button>
-                </div>
-              </Transition.Child>
-              <div className="flex-1 h-0 pt-5 pb-4 overflow-y-auto">
-                <div className="flex-shrink-0 flex items-center px-4">
-                  <img
-                    className="h-8 w-16 inline"
-                    src="/assets/matador.png"
-                    alt="Matador"
-                  />
-                  <h3 className="font-semibold font-xl text-white">Matador</h3>
-                </div>
-                <nav className="mt-5 px-2 space-y-1">{children}</nav>
-              </div>
-              <div className="flex-shrink-0 flex bg-gray-700 p-4">
-                <a href="#" className="flex-shrink-0 group block">
-                  <div className="flex items-center">
-                    <div>
-                      <img
-                        className="h-8 w-16 inline"
-                        src="/assets/matador.png"
-                        alt="Matador"
-                      />
-                      <h3 className="font-semibold font-xl text-white">
-                        Matador
-                      </h3>
-                    </div>
-                  </div>
-                </a>
-              </div>
-            </div>
-          </Transition.Child>
-          <div className="flex-shrink-0 w-14">
-            {/* Force sidebar to shrink to fit close icon */}
-          </div>
-        </Dialog>
-      </Transition.Root>
-
-      {/* Static sidebar for desktop */}
-      <div className="hidden md:flex md:w-64 md:flex-col md:fixed md:inset-y-0">
-        {/* Sidebar component, swap this element with another sidebar if you like */}
-        <div className="flex-1 flex flex-col min-h-0 bg-gray-800">
-          <div className="flex-1 flex flex-col pt-5 pb-4 overflow-y-auto">
-            <div className="flex items-center flex-shrink-0 px-4">
-              <img className="h-8 w-16 inline" src={srcImage} alt={name} />
-              <h3 className="font-semibold font-xl text-white">{name}</h3>
-            </div>
-            <nav className="mt-5 flex-1 px-2 space-y-1">{children}</nav>
-          </div>
-        </div>
-      </div>
-    </>
-  );
-}
