--- app/components/Matador/index.ts
@@ -1,5 +1,0 @@
-export * from "./Card";
-export * from "./JobStatus";
-export * from "./SideBar";
-export * from "./Stat";
-export * from "./Table";
