+++ app/lib/matador/components/job-info/JobInfoPanel.tsx
@@ -14,14 +14,14 @@ export const JobInfoPanel = ({queueName, id, jobName, timestamp, processedOn, fi
             <Divider mb='md' />
 
             <Grid columns={24} mb='md'>
-                <Grid.Col span={12}>
+                <Grid.Col sm={24} xs={24} md={12} lg={12} xl={12}>
                     <Grid>
                         <JobInfo infoName="Queue" infoValue={queueName} />
                         <JobInfo infoName="Job Id" infoValue={id} />
                         <JobInfo infoName="Job Name" infoValue={jobName} />
                     </Grid>
                 </Grid.Col>
-                <Grid.Col span={12}>
+                <Grid.Col sm={24} xs={24} md={12} lg={12} xl={12}>
                     <Grid>
                         <JobInfo infoName="Timestamp" infoValue={timestamp.toISOString()} />
                         { processedOn && <JobInfo infoName="Processed on" infoValue={processedOn.toISOString()} /> }
