+++ app/lib/matador/components/job-result/JobResultPanel.tsx
