+++ app/lib/matador/components/job-result/JobResultPanelProps.ts
