+++ app/lib/matador/components/job-result/JobResultProps.ts
