+++ app/routes/matador.tsx
@@ -18,6 +18,9 @@ export default function MatadorLayout() {
 }
 
 export function CatchBoundary() {
+
+  // FIXME migrate to mantine
+
   const caught = useCatch();
   return (
     <>
