+++ app/routes/matador/queues.tsx
@@ -1,8 +1,8 @@
 import { Grid, Title, Card as MantineCard, Center, Text, Divider } from "@mantine/core";
 import { LoaderFunction } from "@remix-run/node";
 import { NavLink, useLoaderData } from "@remix-run/react";
-import { StatCard } from "../../components/Matador/stat-card";
-import { Link } from "~/components/Matador/ui/Link";
+import { StatCard } from "../../lib/matador/components/stat-card";
+import { Link } from "~/lib/matador/components/ui/Link";
 import { getQueues, getRedisInfo, RedisInfo } from "~/lib/matador/index.server";
 
 type LoaderData = { queues: string[]; serverInfo: RedisInfo };
